requirements.txt
anyio==4.10.0
certifi==2025.8.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-telegram-bot==22.3
sniffio==1.3.1
typing_extensions==4.15.0
